requirements.txt
google-api-python-client==2.100.0
youtube-transcript-api==0.6.1
google-generativeai==0.3.2
python-dotenv==1.0.0
flask==2.3.3
firebase-admin==6.2.0
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.1
pillow==10.0.1
nltk==3.8.1
scikit-learn==1.3.0
